requirements.txt
numpy>=1.19,<1.20
scipy>=1.5,<1.6
pyyaml>=5.3,<5.4
pandas>=1.1,<1.2
pytube>=10.5.3,<10.5.4
requests>=2.25,<2.26
opencv-python>=4.5,<4.6
chess>=1.4,<1.5
matplotlib>=3.3,<3.4
Pillow>=8.1.2,<8.2
autopep8>=1.5,<1.6
setuptools>= 51.0.0,< 51.1.0
wheel>=0.36,<0.37 